requirements.txt
pyodbc
python-dotenv
rich